// app/server/src/db/tables/avwx.drizzle.d.ts
export declare const stations: import("drizzle-orm/mysql-core").MySqlTableWithColumns<{
    name: "stations";
    schema: undefined;
    columns: {
        name: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "name";
            tableName: "stations";
            dataType: "string";
            columnType: "MySqlVarChar";
            data: string;
            driverParam: string | number;
            notNull: false;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: [string, ...string[]];
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        siteId: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "siteId";
            tableName: "stations";
            dataType: "string";
            columnType: "MySqlVarChar";
            data: string;
            driverParam: string | number;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: true;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: [string, ...string[]];
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        lat: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "lat";
            tableName: "stations";
            dataType: "number";
            columnType: "MySqlFloat";
            data: number;
            driverParam: string | number;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        lon: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "lon";
            tableName: "stations";
            dataType: "number";
            columnType: "MySqlFloat";
            data: number;
            driverParam: string | number;
            notNull: true;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        elev_f: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "elev_f";
            tableName: "stations";
            dataType: "number";
            columnType: "MySqlFloat";
            data: number;
            driverParam: string | number;
            notNull: false;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        elev_m: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "elev_m";
            tableName: "stations";
            dataType: "number";
            columnType: "MySqlFloat";
            data: number;
            driverParam: string | number;
            notNull: false;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: undefined;
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        country: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "country";
            tableName: "stations";
            dataType: "string";
            columnType: "MySqlVarChar";
            data: string;
            driverParam: string | number;
            notNull: false;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: [string, ...string[]];
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
        state: import("drizzle-orm/mysql-core").MySqlColumn<{
            name: "state";
            tableName: "stations";
            dataType: "string";
            columnType: "MySqlVarChar";
            data: string;
            driverParam: string | number;
            notNull: false;
            hasDefault: false;
            isPrimaryKey: false;
            isAutoincrement: false;
            hasRuntimeDefault: false;
            enumValues: [string, ...string[]];
            baseColumn: never;
            identity: undefined;
            generated: undefined;
        }, {}, {}>;
    };
    dialect: "mysql";
}>;